fix(auth): parallelize enforce_chat_limit queries and consolidate platform tier fetch
- Run is_byok_active and has_any_oauth_token in parallel via asyncio.gather
  (2 sequential DB round-trips → 1 wall-clock wait)
- Burst guard runs after DB queries succeed so the INCR'd slot isn't
  leaked if a DB connection error propagates
- Move _fetch_platform_tier from users.py to usage_limits.py, rewritten
  to reuse _call_validate_for_user (eliminates duplicate HTTP construction
  and breaks circular deferred import between the two modules)
- Add 15s negative cache on platform tier fetch failure to prevent
  thundering herd against a down service
- Expand effective_is_byok ternary for readability
